cleanup protocol and add tests
